LeSunshine/noise_monitor/requirements.txt
Flask==2.1.1
Flask-SQLAlchemy==2.5.1
Flask-Migrate==3.1.0
python-dotenv==0.20.0
paho-mqtt==1.6.1
SQLAlchemy==1.4.23
Werkzeug==2.0.3
